+++ libre_pvz/src/main.rs
@@ -17,9 +17,6 @@
  */
 
 use bevy::prelude::*;
-use bevy::render::texture::DEFAULT_IMAGE_HANDLE;
-use bevy::sprite::Anchor;
-use libre_pvz::resources::sprite::{Transform, Element};
 use libre_pvz::resources::bevy::{Animation, AnimationLoader};
 
 fn main() {
@@ -46,38 +43,13 @@ fn load_anim(server: Res<AssetServer>, mut commands: Commands) {
 
 fn init_anim(mut ev_anim: EventReader<AssetEvent<Animation>>,
              assets: Res<Assets<Animation>>,
+             mut animation_clips: ResMut<Assets<AnimationClip>>,
              mut commands: Commands) {
     for event in ev_anim.iter() {
         if let AssetEvent::Created { handle } = event {
             let anim = assets.get(handle).unwrap();
-            let parent = commands.spawn_bundle(TransformBundle {
-                local: bevy::prelude::Transform {
-                    scale: Vec3::new(3.0, 3.0, 3.0),
-                    translation: Vec3::new(-120.0, 120.0, 0.0),
-                    ..Default::default()
-                },
-                ..TransformBundle::default()
-            }).id();
             let meta_anim_idle = anim.description.get_meta("anim_idle").unwrap();
-            for (track_id, track) in anim.description.tracks.iter().enumerate() {
-                let frame = &track.frames[meta_anim_idle.start_frame as usize];
-                let mut bundle = SpriteBundle::default();
-                bundle.sprite.anchor = Anchor::TopLeft;
-                for trans in frame.0.iter() {
-                    match trans {
-                        Transform::LoadElement(Element::Text { .. }) => todo!(),
-                        Transform::LoadElement(Element::Image { image }) =>
-                            bundle.texture = anim.images[image].clone(),
-                        Transform::Show(visible) => bundle.visibility.is_visible = *visible,
-                        Transform::Alpha(alpha) => { bundle.sprite.color.set_a(*alpha); }
-                        Transform::Transform(mat) =>
-                            bundle.transform = mat.as_transform_with_z(track_id as f32),
-                    }
-                }
-                if bundle.texture.id == DEFAULT_IMAGE_HANDLE.id { continue; }
-                let child = commands.spawn_bundle(bundle).id();
-                commands.entity(parent).add_child(child);
-            }
+            anim.spawn_on(meta_anim_idle, &mut commands, &mut animation_clips);
         }
     }
 }